requirements.txt
streamlit>=1.42.0
plotly>=5.18.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
statsmodels>=0.14.0
supabase>=2.0.0
requests>=2.31.0
apimoex>=1.3.0
python-dotenv>=1.0.0
pytest>=8.0.0
